@@ -13,36 +13,26 @@ SPARQLWrapper = "*"
 StrEnum = "*"
 pandera = "*"
 pandas = "*"
-python = "^3.9"
-requests = "*"
+python = "^3.10"
 httpx = "*"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.test.dependencies]
 asyncio = "*"
-autopep8 = "*"
-black = { version = "*", allow-prereleases = true }
 coverage = { version = "*", extras = ['toml'] }
-flake8 = "*"
-flake8-bugbear = "*"
-flake8-builtins = "*"
-flake8-comprehensions = "*"
-flake8-debugger = "*"
-flake8-eradicate = "*"
-flake8-mutable = "*"
-flake8-simplify = "*"
-isort = ">= 5"
 mock = "*"
-mypy-ls = "*"
-pre-commit = "*"
-pylint = "*"
-pyls-flake8 = "*"
-pyls-memestra = "*"
 pytest = ">= 6"
 pytest-asyncio = "*"
 pytest-cov = "*"
+pytest-httpserver = "*"
 pytest-integration = "*"
 pytest-profiling = "*"
-pytest-httpserver = "*"
+
+[tool.poetry.group.style.dependencies]
+black = "*"
+ruff = "*"
+
+[tool.poetry.group.dev.dependencies]
+pre-commit = "*"
 
 [tool.poetry.group.docs]
 optional = true
@@ -59,9 +49,89 @@ vcs = "git"
 style = "semver"
 pattern = '^(?P<base>\d+\.\d+\.\d+)(-?((?P<stage>[a-zA-Z]+)\.?(?P<revision>\d+)?))?$'
 
-[tool.isort]
-profile = "black"
-line_length = 100
+[tool.ruff]
+select = [
+  "A", # flake8-builtins
+  "ANN", # flake8-annotations
+  "ARG", # flake8-unused-arguments
+  "B", # flake8-bugbear
+  "C4", # flake8-comprehensions
+  "C90", # mccabe
+  "COM", # flake8-commas
+  "DTZ", # flake8-datetimez
+  "E", "W", # pycodestyle
+  "F", # Pyflakes
+  "FLY", # flynt
+  "FURB", # refurb
+  "I", # isort
+  "N", # pep8-nameing
+  "NPY", # numpy specific rules
+  "PERF", # Perflint
+  "PIE", # flake8-pie
+  "RUF", # Ruff specific rules
+  "S", # flake8-bandit
+  "SIM", # flake8-simplify
+  "T20", # flake8-print
+  "TCH001", # flake8-type-checking
+  "TRY", # tryceratops
+  "UP", # pyupgrade
+  "YTT", # flake8-2020
+]
+
+ignore = ['S101', 'COM812', 'ANN101', 'ANN102', 'ANN401', 'TRY003']
+
+# Allow autofix for all enabled rules (when `--fix`) is provided.
+fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
+unfixable = []
+
+# Exclude a variety of commonly ignored directories.
+exclude = [
+    ".bzr",
+    ".direnv",
+    ".eggs",
+    ".git",
+    ".git-rewrite",
+    ".hg",
+    ".mypy_cache",
+    ".nox",
+    ".pants.d",
+    ".pytype",
+    ".ruff_cache",
+    ".svn",
+    ".tox",
+    ".venv",
+    "__pypackages__",
+    "_build",
+    "buck-out",
+    "build",
+    "dist",
+    "node_modules",
+    "venv",
+]
+
+# Same as Black.
+line-length = 100
+
+# Allow unused variables when underscore-prefixed.
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+# Assume Python 3.10
+target-version = "py310"
+
+[tool.ruff.per-file-ignores]
+"docs/conf.py" = ["A001"]
+"tests/test*.py" = ["ANN201"]
+
+[tool.ruff.pep8-naming]
+# Allow Pydantic's `@validator` decorator to trigger class method treatment.
+classmethod-decorators = ["pandera.dataframe_check"]
+
+[tool.ruff.flake8-import-conventions]
+
+[tool.ruff.flake8-import-conventions.aliases]
+# Declare the default aliases.
+pandas = "pd"
+pandera = "pa"
 
 [tool.black]
 line-length = 100
