@@ -1,12 +1,14 @@
+from __future__ import annotations
+
 import datetime
 from decimal import Decimal
-from typing import Optional
 
 import numpy as np
 import pandas as pd
 import pandera as pa
 import pytest
 from pandas.testing import assert_frame_equal
+from pytest_httpserver import HTTPServer
 
 from cimsparql import type_mapper
 from cimsparql.data_models import JsonSchemaOut
@@ -14,7 +16,9 @@ from cimsparql.graphdb import RestApi
 from cimsparql.model import ServiceConfig
 
 
-def init_triple_store_server(httpserver, sparql_result: Optional[dict] = None) -> ServiceConfig:
+def init_triple_store_server(
+    httpserver: HTTPServer, sparql_result: dict | None = None
+) -> ServiceConfig:
     """
     Create a triple store server that returns sparql_result_data when a call is made
     """
@@ -34,13 +38,13 @@ def test_python_type_map_bool():
     assert not type_mapper.XSD_TYPE_MAP["boolean"]("FALSE")
 
 
-def test_get_map_empty_pandas(httpserver):
+def test_get_map_empty_pandas(httpserver: HTTPServer):
     service_cfg = init_triple_store_server(httpserver)
     mapper = type_mapper.TypeMapper(service_cfg)
     assert mapper.get_map() == {}
 
 
-def test_map_data_types(httpserver):
+def test_map_data_types(httpserver: HTTPServer):
     results = {
         "sparql_type": ["http://c#Degrees", "http://c#Status", "http://c#Amount"],
         "range": ["http://x#Float", "http://x#Bool", "http://x#Integer"],
@@ -89,7 +93,10 @@ def test_map_data_types(httpserver):
         ("boolean", ("1", True)),
         ("boolean", ("0", False)),
         ("date", ("2021-12-14", pd.Timestamp("2021-12-14 00:00:00"))),
-        ("dateTime", ("2002-10-14T12:00:00", datetime.datetime(2002, 10, 14, 12, 0, 0))),
+        (
+            "dateTime",
+            ("2002-10-14T12:00:00", datetime.datetime(2002, 10, 14, 12, 0, 0)),  # noqa DTZ001
+        ),
         ("dateTime", ("2002-10-14T12:00:00Z", pd.Timestamp("2002-10-14 12:00:00+0000", tz="UTC"))),
         ("duration", ("P365D", datetime.timedelta(365))),
         ("duration", ("-P365D", datetime.timedelta(-365))),
@@ -115,7 +122,7 @@ def test_xsd_types(dtype: str, test: tuple):
         ("16", True),
     ],
 )
-def test_have_cim_version(httpserver, cim_version: int, have_cim_version: bool):
+def test_have_cim_version(httpserver: HTTPServer, cim_version: int, have_cim_version: bool):
     service_cfg = init_triple_store_server(httpserver)
     tm = type_mapper.TypeMapper(service_cfg)
     assert tm.have_cim_version(cim_version) == have_cim_version
@@ -132,7 +139,7 @@ class FloatSchema(JsonSchemaOut):
         ({"float_col": ["1.0", None]}, {"float_col": [1.0, np.nan]}),
     ],
 )
-def test_coerce_missing_type(httpserver, in_data: dict, out_data: dict):
+def test_coerce_missing_type(httpserver: HTTPServer, in_data: dict, out_data: dict):
     client = init_triple_store_server(httpserver)
     in_df, out_df = pd.DataFrame(in_data), pd.DataFrame(out_data)
     mapper = type_mapper.TypeMapper(client)
@@ -142,7 +149,7 @@ def test_coerce_missing_type(httpserver, in_data: dict, out_data: dict):
     pd.testing.assert_frame_equal(df, out_df)
 
 
-def test_coerce_float_from_literal(httpserver):
+def test_coerce_float_from_literal(httpserver: HTTPServer):
     client = init_triple_store_server(httpserver)
     mapper = type_mapper.TypeMapper(client)
 
