@@ -2,7 +2,7 @@ import pandas as pd
 
 
 class ValueMapper:
-    def map(self, df: pd.DataFrame) -> pd.DataFrame:
+    def map(self, df: pd.DataFrame) -> pd.DataFrame:  # noqa A003
         return df
 
 
@@ -15,7 +15,7 @@ class MridMapper(ValueMapper):
     def __init__(self, uri_regex: str = "^([^#_|^#]+)(#_|#)") -> None:
         self.uri_regex = uri_regex
 
-    def map(self, df: pd.DataFrame) -> pd.DataFrame:
+    def map(self, df: pd.DataFrame) -> pd.DataFrame:  # noqa A003
         for col in filter(lambda col: "mrid" in col, df.columns):
             df[col] = df[col].str.replace(self.uri_regex, "", regex=True)
         return df
